Fix imports by renaming conflicting packages
