+++ wf_logging/__init__.py
@@ -1,0 +1,5 @@
+"""Lightweight logging utilities used by WiFi Fixer Tool."""
+
+from .logger import Logger
+
+__all__ = ["Logger"]
